home-page: abstract the fetching/state hook
